% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_symptoms.r
\name{aggregate_symptoms}
\alias{aggregate_symptoms}
\title{Aggregate symptoms in symptom table to work out ILI etc.}
\usage{
aggregate_symptoms(symptoms, ili.sudden.unknown = 1)
}
\arguments{
\item{symptoms}{}
}
\value{
a data table with added columns
}
\description{
Aggregate symptoms in symptom table to work out ILI etc.
}
\author{
seb
}

