% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouts.r
\name{bouts_of_illness}
\alias{bouts_of_illness}
\title{Extract bouts of illness from a flusurvey data table}
\usage{
bouts_of_illness(x, progress = TRUE)
}
\arguments{
\item{x}{the data to extract bouts from}

\item{progress}{whether to display a progress bar (default: TRUE)}
}
\value{
a data table of bouts of illness
}
\description{
Extract bouts of illness from a flusurvey data table
}
\author{
seb
}
