\name{flusurvey-package}
\alias{flusurvey-package}
\alias{flusurvey}
\docType{package}
\title{
\packageTitle{flusurvey}
}
\description{
\packageDescription{flusurvey}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{flusurvey}
\packageIndices{flusurvey}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{flusurvey}

Maintainer: \packageMaintainer{flusurvey}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
