% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data.r
\name{read_data}
\alias{read_data}
\title{This reads in flusurvey data for a given year}
\usage{
read_data(files, year, ...)
}
\arguments{
\item{files}{a (named!) list of files. The names stand for the surveys (e.g., "symptom", "background", "contact", etc.)}

\item{year}{the year from which the data come. If not given, will try to guess it from the headers in the .csv files}

\item{...}{options to be passed to \code{read.csv}}
}
\value{
a list of data tables with the data
}
\description{
This reads different surveys and returns a list of data tables
containing the surveys. It does minimal cleaning and conversion
into R data formats. The csv files that should be provided can be
obtained via SQL dump,  e.g.
\\f ','
\\a
\\t
\\o intake_16.csv
SELECT * FROM pollster_results_intake;
\\o
\\q
}
\details{
and similarly for weekly surveys
}
\author{
seb
}

