% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouts.r
\name{bouts_of_illness}
\alias{bouts_of_illness}
\title{Extract bouts of illness from a flusurvey data table}
\usage{
bouts_of_illness(x, symptomatic.only = TRUE, as.data.frame = TRUE)
}
\arguments{
\item{x}{the data to extract bouts from}

\item{symptomatic.only}{whether to only include symptomatic periods}
}
\value{
a data table of bouts of illness
}
\description{
Extract bouts of illness from a flusurvey data table
}
\author{
seb
}
