% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.r
\name{merge_data}
\alias{merge_data}
\title{Clean and merge flusurvey data tables}
\usage{
merge_data(data, clean = c("remove.first", "remove.bad.symptom.dates",
  "guess.start.dates", "limit.season", "remove.postcodes", "create.numeric.id",
  "n.reports"), min.reports = 3)
}
\arguments{
\item{data}{the data to merge and clean, usually the result of \code{read_data}}

\item{clean}{cleaning options, NULL for no cleaning, otherwise a vector of cleans to perform (by default all):
- 'remove.first', whether to remove everyone's first report,
- 'remove.bad.symptom.dates', whether to remove rows with the symptom end date before the symptom start date, or with symptom dates outside the reporting dates
- 'guess.start.dates', whether to guess the symptom start dates when the person couldn't remember (as a day of the report)
- 'limit.season', whether to limit a flu season to November -> April
- 'remove.postcodes', whether to limit a flu season to November -> April
- 'create.numeric.id', whether to create a numeric id for every user,
- 'n.reports', whether to exclude those with fewer than \code{min.reports} reports}

\item{min.reports}{minimum number of reports per user (ignored if 'min.reports' is not given as a cleaning option)}
}
\value{
a rolling-joined data table
}
\description{
Clean and merge flusurvey data tables
}
\author{
seb
}

